+++ include/devices/dragino.h
@@ -44,7 +44,7 @@ class Dragino: public Device {
         void setGps();
         gps_fix getGpsFix();
         void setLora();
-        void sendLora(uint8_t LoRaPacketData[], uint8_t LoRaPacketDataSize);
+        // void sendLora(uint8_t LoRaPacketData[], uint8_t LoRaPacketDataSize);
 };
 
 #endif
