+++ include/devices/rak7200.h
@@ -78,7 +78,6 @@
 
 #define SX1276_RegVersion                 0x42
 
-const unsigned TX_INTERVAL = 15; // Transmit every 15 seconds
 
 extern class Rak7200 dev;
 
@@ -89,25 +88,25 @@ class Rak7200: public Device {
         void setGps();
         gps_fix getGpsFix();
         void setLora();
-        void sendLora(uint8_t LoRaPacketData[], uint8_t LoRaPacketDataSize);
+        // void sendLora(uint8_t LoRaPacketData[], uint8_t LoRaPacketDataSize);
 
         // void do_send(osjob_t *j);
-        uint8_t* getLoRaPacketData();
-        uint8_t  getLoRaPacketDataSize();
-        void resetLoRaPacketDataSize();
-        osjob_t* getSendjob();
+        // uint8_t* getLoRaPacketData();
+        // uint8_t  getLoRaPacketDataSize();
+        // void resetLoRaPacketDataSize();
+        // osjob_t* getSendjob();
 
     private:
         bool GNSS_probe();
         HardwareSerial* _GNSS = NULL;
 
         bool _has_SX1276;
-        void LmicInit();
+        // void LmicInit();
         // void onEvent(ev_t ev);
 
-        osjob_t _sendjob;
-        uint8_t _loRaPacketData[24];
-        uint8_t _loRaPacketDataSize;
+        // osjob_t _sendjob;
+        // uint8_t _loRaPacketData[24];
+        // uint8_t _loRaPacketDataSize;
