+++ src/devices/rak7200.cpp
@@ -37,23 +37,7 @@
 #include <devices\rak7200.h>
 #include <SPI.h>
 
-const lmic_pinmap lmic_pins = {
-        .nss = S7xx_SX127x_NSS,
-        .rxtx = S7xx_SX127x_ANTENNA_SWITCH_RXTX,
-        .rst = S7xx_SX127x_NRESET,
-        .dio = {S7xx_SX127x_DIO0, S7xx_SX127x_DIO1, S7xx_SX127x_DIO2},
-        .rxtx_rx_active = 1,
-        .rssi_cal = 10,
-        .spi_freq = 1000000
-};
 
-void printHex2(unsigned v) {
-    v &= 0xff;
-    if (v < 16) {
-        Serial.print('0');
-    }
-    Serial.print(v, HEX);
-}
 
 Rak7200::Rak7200(){
     _GNSS = new HardwareSerial(S7xG_CXD5603_UART_RX, S7xG_CXD5603_UART_TX);
@@ -226,189 +210,34 @@ void Rak7200::setLora(){
     LmicInit();
 }
 
-void Rak7200::sendLora(uint8_t LoRaPacketData[], uint8_t LoRaPacketDataSize){
-    memcpy_P(_loRaPacketData, LoRaPacketData, LoRaPacketDataSize);
-    _loRaPacketDataSize = LoRaPacketDataSize;
-}
+// LMIC library expects pinmap as global constant
+const lmic_pinmap lmic_pins = {
+        .nss = S7xx_SX127x_NSS,
+        .rxtx = S7xx_SX127x_ANTENNA_SWITCH_RXTX,
+        .rst = S7xx_SX127x_NRESET,
+        .dio = {S7xx_SX127x_DIO0, S7xx_SX127x_DIO1, S7xx_SX127x_DIO2},
+        .rxtx_rx_active = 1,
+        .rssi_cal = 10,
+        .spi_freq = 1000000
+};
+
+// void Rak7200::sendLora(uint8_t LoRaPacketData[], uint8_t LoRaPacketDataSize){
+//     memcpy_P(_loRaPacketData, LoRaPacketData, LoRaPacketDataSize);
+//     _loRaPacketDataSize = LoRaPacketDataSize;
+// }
 
 
-uint8_t* Rak7200::getLoRaPacketData(){
-    return _loRaPacketData;
-}
-uint8_t  Rak7200::getLoRaPacketDataSize(){
-    return _loRaPacketDataSize;
-}
+// uint8_t* Rak7200::getLoRaPacketData(){
+//     return _loRaPacketData;
+// }
+// uint8_t  Rak7200::getLoRaPacketDataSize(){
+//     return _loRaPacketDataSize;
+// }
 
-void  Rak7200::resetLoRaPacketDataSize(){
-    _loRaPacketDataSize = 0;
-}
+// void  Rak7200::resetLoRaPacketDataSize(){
+//     _loRaPacketDataSize = 0;
+// }
 
-osjob_t* Rak7200::getSendjob(){
-    return &_sendjob;
-}
-
-
-void do_send(osjob_t *j) {
-    // Check if there is not a current TX/RX job running
-    if (LMIC.opmode & OP_TXRXPEND) {
-        Serial.println(F("OP_TXRXPEND, not sending"));
-    }
-    else {
-        LMIC_setTxData2(1, dev.getLoRaPacketData(), dev.getLoRaPacketDataSize(), 0);
-        dev.resetLoRaPacketDataSize();
-        Serial.println(F("Packet queued"));
-    }
-    // Next TX is scheduled after TX_COMPLETE event.
-}
-
-void Rak7200::LmicInit(){
-    // LMIC init
-    os_init();
-    // Reset the MAC state. Session and pending data transfers will be discarded.
-    LMIC_reset();
-
-    // allow much more clock error than the X/1000 default. See:
-    // https://github.com/mcci-catena/arduino-lorawan/issues/74#issuecomment-462171974
-    // https://github.com/mcci-catena/arduino-lmic/commit/42da75b56#diff-16d75524a9920f5d043fe731a27cf85aL633
-    // the X/1000 means an error rate of 0.1%; the above issue discusses using values up to 10%.
-    // so, values from 10 (10% error, the most lax) to 1000 (0.1% error, the most strict) can be used.
-    //LMIC_setClockError(1 * MAX_CLOCK_ERROR / 40);
-    LMIC_setClockError(MAX_CLOCK_ERROR *20 / 100);
-
-    LMIC_setLinkCheckMode(0);
-    LMIC_setDrTxpow(DR_SF12, 14);
-    //LMIC_setupBand(BAND_MILLI,13,10);
-    //LMIC_selectSubBand(1);
-    Serial.println("Radio Initialized");
-
-    // Start job (sending automatically starts OTAA too)
-    do_send(&_sendjob);
-}
-
-
-
-void onEvent(ev_t ev) {
-    Serial.print(os_getTime());
-    Serial.print(": ");
-    switch (ev) {
-        case EV_SCAN_TIMEOUT:
-            Serial.println(F("EV_SCAN_TIMEOUT"));
-            break;
-        case EV_BEACON_FOUND:
-            Serial.println(F("EV_BEACON_FOUND"));
-            break;
-        case EV_BEACON_MISSED:
-            Serial.println(F("EV_BEACON_MISSED"));
-            break;
-        case EV_BEACON_TRACKED:
-            Serial.println(F("EV_BEACON_TRACKED"));
-            break;
-        case EV_JOINING:
-            Serial.println(F("EV_JOINING"));
-            break;
-        case EV_JOINED:
-            Serial.println(F("EV_JOINED"));
-            {
-                u4_t netid = 0;
-                devaddr_t devaddr = 0;
-                u1_t nwkKey[16];
-                u1_t artKey[16];
-                LMIC_getSessionKeys(&netid, &devaddr, nwkKey, artKey);
-                Serial.print("netid: ");
-                Serial.println(netid, DEC);
-                Serial.print("devaddr: ");
-                Serial.println(__builtin_bswap32(devaddr), HEX);
-                Serial.print("AppSKey: ");
-                for (size_t i = 0; i < sizeof(artKey); ++i) {
-                    if (i != 0) {
-                        Serial.print("-");
-                    }
-                    printHex2(artKey[i]);
-                }
-                Serial.println("");
-                Serial.print("NwkSKey: ");
-                for (size_t i = 0; i < sizeof(nwkKey); ++i) {
-                    if (i != 0) {
-                        Serial.print("-");
-                    }
-                    printHex2(nwkKey[i]);
-                }
-                Serial.println();
-            }
-            // Disable link check validation (automatically enabled
-            // during join, but because slow data rates change max TX
-            // size, we don't use it in this example.
-            LMIC_setLinkCheckMode(0);
-            break;
-            /*
-            || This event is defined but not used in the code. No
-            || point in wasting codespace on it.
-            ||
-            || case EV_RFU1:
-            ||     Serial.println(F("EV_RFU1"));
-            ||     break;
-            */
-        case EV_JOIN_FAILED:
-            Serial.println(F("EV_JOIN_FAILED"));
-            break;
-        case EV_REJOIN_FAILED:
-            Serial.println(F("EV_REJOIN_FAILED"));
-            break;
-            break;
-        case EV_TXCOMPLETE:
-            Serial.println(F("EV_TXCOMPLETE (includes waiting for RX windows)"));
-            if (LMIC.txrxFlags & TXRX_ACK) {
-                Serial.println(F("Received ack"));
-            }
-            if (LMIC.dataLen) {
-                Serial.println(F("Received "));
-                Serial.println(LMIC.dataLen);
-                Serial.println(F(" bytes of payload"));
-            }
-            // Schedule next transmission
-            os_setTimedCallback(dev.getSendjob(), os_getTime() + sec2osticks(TX_INTERVAL), do_send);
-            break;
-        case EV_LOST_TSYNC:
-            Serial.println(F("EV_LOST_TSYNC"));
-            break;
-        case EV_RESET:
-            Serial.println(F("EV_RESET"));
-            break;
-        case EV_RXCOMPLETE:
-            // data received in ping slot
-            Serial.println(F("EV_RXCOMPLETE"));
-            break;
-        case EV_LINK_DEAD:
-            Serial.println(F("EV_LINK_DEAD"));
-            break;
-        case EV_LINK_ALIVE:
-            Serial.println(F("EV_LINK_ALIVE"));
-            break;
-            /*
-            || This event is defined but not used in the code. No
-            || point in wasting codespace on it.
-            ||
-            || case EV_SCAN_FOUND:
-            ||    Serial.println(F("EV_SCAN_FOUND"));
-            ||    break;
-            */
-        case EV_TXSTART:
-            Serial.println(F("EV_TXSTART"));
-            break;
-        case EV_TXCANCELED:
-            Serial.println(F("EV_TXCANCELED"));
-            break;
-        case EV_RXSTART:
-            /* do not print anything -- it wrecks timing */
-            break;
-        case EV_JOIN_TXCOMPLETE:
-            Serial.println(F("EV_JOIN_TXCOMPLETE: no JoinAccept"));
-            break;
-
-        default:
-            Serial.print(F("Unknown event: "));
-            Serial.println((unsigned) ev);
-            break;
-    }
-}
-
+// osjob_t* Rak7200::getSendjob(){
+//     return &_sendjob;
+// }
